refactor: 불필요한 객체 제거
- NoteResult
- NoteService -> NoteAPIService로 대체
- Source 폴더 내부에 NoteAPIService로 extension 메서드 구현
