--- Projects/Domain/Note/Interface/Sources/Model/NoteResult.swift
@@ -1,13 +1,0 @@
-//
-//  NoteResult.swift
-//  DomainNoteInterface
-//
-//  Created by Derrick kim on 8/21/24.
-//
-
-import Foundation
-
-public enum NoteResult {
-    case success
-    case failure(NoteError)
-}
