+++ Projects/Feature/Main/Example/Sources/SceneDelegate.swift
@@ -170,8 +170,8 @@ extension SceneDelegate: PostNoteViewControllerDelegate, SearchSongViewControlle
     }
 
     func postNoteDependencies(artistID: Int) -> PostNoteViewModel {
-        @Injected(.noteService) var noteService: NoteServiceInterface
-        let postNoteUseCase: PostNoteUseCaseInterface = PostNoteUseCase(noteService: noteService)
+        @Injected(.noteAPIService) var noteAPIService: NoteAPIServiceInterface
+        let postNoteUseCase: PostNoteUseCaseInterface = PostNoteUseCase(noteAPIService: noteAPIService)
         let viewModel = PostNoteViewModel(postNoteUseCase: postNoteUseCase, artistID: artistID)
 
         return viewModel
@@ -187,13 +187,13 @@ extension SceneDelegate: PostNoteViewControllerDelegate, SearchSongViewControlle
     }
 
     func pushSearchSongViewController(artistID: Int) {
-        @Injected(.noteService) var noteService: NoteServiceInterface
+        @Injected(.noteAPIService) var noteAPIService: NoteAPIServiceInterface
         @Injected(.songPaginationService) var songPaginationService: SongPaginationServiceInterface
 
         songPaginationService.resetPagination()
 
         let searchSongUseCase = SearchSongUseCase(
-            noteService: noteService,
+            noteAPIService: noteAPIService,
             songPaginationService: songPaginationService
         )
         let searchSongViewModel = SearchSongViewModel(
